refactor(schema): split UpdateSchema into focused modules
